Use SB's formatting

+++ no-framework/lib/framework.js
@@ -1,20 +1,20 @@
 #!/usr/env/bin node
 
-import ejs from "ejs";
-import fs from "fs";
-import glob from "fast-glob";
-import liveServer from "live-server";
-import path from "path";
-import prettier from "prettier";
+import ejs from 'ejs';
+import fs from 'fs';
+import glob from 'fast-glob';
+import liveServer from 'live-server';
+import path from 'path';
+import prettier from 'prettier';
 
 /* ----- Constants ----- */
 
-const IS_DEV = process.argv.includes("dev");
-const SRC_DIR = path.join(process.cwd(), "src");
-const DIST_DIR = path.join(process.cwd(), "dist");
-const PAGES_DIR = path.join(SRC_DIR, "content/pages");
-const LAYOUTS_DIR = path.join(SRC_DIR, "layouts");
-const COMPONENTS_DIR = path.join(SRC_DIR, "components");
+const IS_DEV = process.argv.includes('dev');
+const SRC_DIR = path.join(process.cwd(), 'src');
+const DIST_DIR = path.join(process.cwd(), 'dist');
+const PAGES_DIR = path.join(SRC_DIR, 'content/pages');
+const LAYOUTS_DIR = path.join(SRC_DIR, 'layouts');
+const COMPONENTS_DIR = path.join(SRC_DIR, 'components');
 
 /* ----- References ----- */
 
@@ -25,8 +25,7 @@ let components = {};
 /* ----- Setup ----- */
 
 // Remove existing dist directory and built files
-if (fs.existsSync(DIST_DIR))
-  fs.rmSync(DIST_DIR, { recursive: true, force: true });
+if (fs.existsSync(DIST_DIR)) fs.rmSync(DIST_DIR, { recursive: true, force: true });
 // Create new dist directory
 fs.mkdirSync(DIST_DIR);
 
@@ -40,16 +39,11 @@ fs.mkdirSync(DIST_DIR);
  * @param {string} watchDir - Absolute path to directory being watched
  */
 function updateSite(fileChanged, watchDir) {
-  console.log(
-    `[Source Change] ${path.relative(
-      process.cwd(),
-      path.join(watchDir, fileChanged)
-    )}`
-  );
-  // If layout file or component is changed, rebuild the entire site
-  if ([LAYOUTS_DIR, COMPONENTS_DIR].includes(watchDir)) return buildSite();
-  // Otherwise, just rebuild the page that changed
-  buildPage(fileChanged);
+    console.log(`[Source Change] ${path.relative(process.cwd(), path.join(watchDir, fileChanged))}`);
+    // If layout file or component is changed, rebuild the entire site
+    if ([LAYOUTS_DIR, COMPONENTS_DIR].includes(watchDir)) return buildSite();
+    // Otherwise, just rebuild the page that changed
+    buildPage(fileChanged);
 }
 
 /**
@@ -57,16 +51,14 @@ function updateSite(fileChanged, watchDir) {
  * file is changed.
  */
 function updateLayoutsRef() {
-  const layoutFiles = glob.sync("**/*.ejs", { cwd: LAYOUTS_DIR });
-  layouts = Object.fromEntries(
-    layoutFiles.map((filePath) => {
-      const layoutName = path.basename(filePath, ".ejs");
-      const layout = fs
-        .readFileSync(path.join(LAYOUTS_DIR, filePath), "utf-8")
-        .toString();
-      return [layoutName, layout];
-    })
-  );
+    const layoutFiles = glob.sync('**/*.ejs', { cwd: LAYOUTS_DIR });
+    layouts = Object.fromEntries(
+        layoutFiles.map((filePath) => {
+            const layoutName = path.basename(filePath, '.ejs');
+            const layout = fs.readFileSync(path.join(LAYOUTS_DIR, filePath), 'utf-8').toString();
+            return [layoutName, layout];
+        })
+    );
 }
 
 /**
@@ -74,16 +66,14 @@ function updateLayoutsRef() {
  * file is changed.
  */
 function updateComponentsRef() {
-  const componentFiles = glob.sync("**/*.ejs", { cwd: COMPONENTS_DIR });
-  components = Object.fromEntries(
-    componentFiles.map((filePath) => {
-      const componentName = path.basename(filePath, ".ejs");
-      const component = fs
-        .readFileSync(path.join(COMPONENTS_DIR, filePath), "utf-8")
-        .toString();
-      return [componentName, component];
-    })
-  );
+    const componentFiles = glob.sync('**/*.ejs', { cwd: COMPONENTS_DIR });
+    components = Object.fromEntries(
+        componentFiles.map((filePath) => {
+            const componentName = path.basename(filePath, '.ejs');
+            const component = fs.readFileSync(path.join(COMPONENTS_DIR, filePath), 'utf-8').toString();
+            return [componentName, component];
+        })
+    );
 }
 
 /* ----- Renderers----- */
@@ -96,7 +86,7 @@ function updateComponentsRef() {
  * @returns
  */
 function renderComponent(props) {
-  return ejs.render(components[props.type], { ...props });
+    return ejs.render(components[props.type], { ...props });
 }
 
 /* ----- Builders ----- */
@@ -106,15 +96,15 @@ function renderComponent(props) {
  * whenever a source file is changed.
  */
 function buildSite() {
-  // Rebuild global references
-  updateLayoutsRef();
-  updateComponentsRef();
-  // Get page files
-  const pageFiles = glob.sync("**/*.json", { cwd: PAGES_DIR });
-  // Build each page
-  pageFiles.forEach(buildPage);
-  // Provide feedback
-  console.log(`Built ${pageFiles.length} pages`);
+    // Rebuild global references
+    updateLayoutsRef();
+    updateComponentsRef();
+    // Get page files
+    const pageFiles = glob.sync('**/*.json', { cwd: PAGES_DIR });
+    // Build each page
+    pageFiles.forEach(buildPage);
+    // Provide feedback
+    console.log(`Built ${pageFiles.length} pages`);
 }
 
 /**
@@ -124,67 +114,59 @@ function buildSite() {
  * @param {string} relSrcFilePath - Path to page file, relative to PAGES_DIR
  */
 function buildPage(relSrcFilePath) {
-  const absSrcFilePath = path.join(PAGES_DIR, relSrcFilePath);
-  // Get and set urlPath on page from file path
-  const urlPath = relSrcFilePath
-    .replace(/\.json$/, "/index.html")
-    .replace(/\/index\/index\.html$/, "/index.html")
-    .replace(/^index\/index\.html$/, "index.html");
-  // Determine destination file path
-  const distFilePath = path.join(DIST_DIR, urlPath);
-  // If the source file doesn't exist, first delete the file in the dist if it
-  // exists, then return
-  if (!fs.existsSync(absSrcFilePath)) {
-    if (fs.existsSync(distFilePath)) fs.rmSync(distFilePath);
-    return;
-  }
-  // Parse page and retrieve layout
-  const rawPageContent = fs.readFileSync(absSrcFilePath, "utf-8").toString();
-  const page = JSON.parse(rawPageContent);
-  const layout = layouts[page.layout];
-  // Escape if layout doesn't exist
-  if (!layout) {
-    console.log(
-      `[Error] Layout "${page.layout}" not found for page "${relSrcFilePath}"`
-    );
-    return;
-  }
-  // Add meta information for the page
-  page._meta = {
-    // `id` is path from root of project (for inline editing)
-    id: path.relative(process.cwd(), absSrcFilePath),
-  };
-  // Create directory if it doesn't exist
-  const dirPath = path.dirname(distFilePath);
-  if (!fs.existsSync(dirPath)) fs.mkdirSync(dirPath, { recursive: true });
-  // Run page through EJS layout and write to file
-  const html = ejs.render(layout, { page, component: renderComponent });
-  fs.writeFileSync(distFilePath, prettier.format(html, { parser: "html" }));
+    const absSrcFilePath = path.join(PAGES_DIR, relSrcFilePath);
+    // Get and set urlPath on page from file path
+    const urlPath = relSrcFilePath
+        .replace(/\.json$/, '/index.html')
+        .replace(/\/index\/index\.html$/, '/index.html')
+        .replace(/^index\/index\.html$/, 'index.html');
+    // Determine destination file path
+    const distFilePath = path.join(DIST_DIR, urlPath);
+    // If the source file doesn't exist, first delete the file in the dist if it
+    // exists, then return
+    if (!fs.existsSync(absSrcFilePath)) {
+        if (fs.existsSync(distFilePath)) fs.rmSync(distFilePath);
+        return;
+    }
+    // Parse page and retrieve layout
+    const rawPageContent = fs.readFileSync(absSrcFilePath, 'utf-8').toString();
+    const page = JSON.parse(rawPageContent);
+    const layout = layouts[page.layout];
+    // Escape if layout doesn't exist
+    if (!layout) {
+        console.log(`[Error] Layout "${page.layout}" not found for page "${relSrcFilePath}"`);
+        return;
+    }
+    // Add meta information for the page
+    page._meta = {
+        // `id` is path from root of project (for inline editing)
+        id: path.relative(process.cwd(), absSrcFilePath)
+    };
+    // Create directory if it doesn't exist
+    const dirPath = path.dirname(distFilePath);
+    if (!fs.existsSync(dirPath)) fs.mkdirSync(dirPath, { recursive: true });
+    // Run page through EJS layout and write to file
+    const html = ejs.render(layout, { page, component: renderComponent });
+    fs.writeFileSync(distFilePath, prettier.format(html, { parser: 'html' }));
 }
 
 /* ----- Watchers / Callers ----- */
 
 if (IS_DEV) {
-  // Watch for changes to content source files and rebuild
-  fs.watch(PAGES_DIR, { recursive: true }, (_, filename) =>
-    updateSite(filename, PAGES_DIR)
-  );
-  // Watch for changes to layout files and rebuild
-  fs.watch(LAYOUTS_DIR, { recursive: true }, (_, filename) =>
-    updateSite(filename, LAYOUTS_DIR)
-  );
-  // Watch for changes to component files and rebuild
-  fs.watch(COMPONENTS_DIR, { recursive: true }, (_, filename) =>
-    updateSite(filename, COMPONENTS_DIR)
-  );
-  // Start development server, which serves from and watches for changes in the
-  // dist directory
-  liveServer.start({
-    port: 3000,
-    root: path.relative(process.cwd(), DIST_DIR),
-    open: false,
-    host: "localhost",
-  });
+    // Watch for changes to content source files and rebuild
+    fs.watch(PAGES_DIR, { recursive: true }, (_, filename) => updateSite(filename, PAGES_DIR));
+    // Watch for changes to layout files and rebuild
+    fs.watch(LAYOUTS_DIR, { recursive: true }, (_, filename) => updateSite(filename, LAYOUTS_DIR));
+    // Watch for changes to component files and rebuild
+    fs.watch(COMPONENTS_DIR, { recursive: true }, (_, filename) => updateSite(filename, COMPONENTS_DIR));
+    // Start development server, which serves from and watches for changes in the
+    // dist directory
+    liveServer.start({
+        port: 3000,
+        root: path.relative(process.cwd(), DIST_DIR),
+        open: false,
+        host: 'localhost'
+    });
 }
 
 // Do the initial build
